feat: Add drag and drop functionality to front end
Drag and drop into sections is made possible by this commit, although there is a bug that remains that doubles the item on most drops. I have left evidence of debugging to show my thought process (passing cards through a set object still yields duplicates), but was unable to solve the issue in the time available. Dropping into a card is not complete by this commit but there is a variable that detects that a card is the drop target.